#372: Moved the AbilityKillerDamageIncreaseParser and its tests to the model folder
The class was not modified yet.
