+++ src/app/ffbe/model/effects/abilities/ability-killer-damage-increase.parser.spec.ts
@@ -1,4 +1,4 @@
-import {AbilityEffectParserFactory} from './ability-effect-parser.factory';
+import {AbilityEffectParserFactory} from '../../../mappers/effects/abilities/ability-effect-parser.factory';
 
 describe('AbilityKillerDamageIncreaseParser', () => {
   it('should parse physical beast killer for caster', () => {
