+++ src/app/ffbe/model/effects/abilities/ability-killer-damage-increase.parser.ts
@@ -1,8 +1,8 @@
-import {EffectParser} from '../effect-parser';
-import {Skill} from '../../../model/skill.model';
+import {EffectParser} from '../../../mappers/effects/effect-parser';
+import {Skill} from '../../skill.model';
 import {FFBE_MONSTER_TYPES} from '../../../ffbe.constants';
 import {FfbeUtils} from '../../../utils/ffbe-utils';
-import {HTML_LINE_RETURN} from '../skill-effects.mapper';
+import {HTML_LINE_RETURN} from '../../../mappers/effects/skill-effects.mapper';
 
 export class AbilityKillerDamageIncreaseParser extends EffectParser {
   private target: string;
